Create a StackerCommand to share common parameters between stacker commands
The way argparse handles arguments that are added to a parent parser and
meant to be shared by the subparsers is confusing. The StackerCommand
makes it so each subclass will have the shared parameters correctly
added to the subparser.

Some of this weirdness is outlined here:

https://bugs.python.org/issue9540
https://bugs.python.org/issue9253

And more is explained in the StackerCommand docstring
